A long, long overdue integration of all our recent hard work out to the open-source project. I will now generate and post a new snapshot build.
What's new:
 * Functions and Predicates each implement eq()/hC()/tS() and are serializable
 * Functions: toHashCode(), toStringFunction() (replaces TO_STRING)
 * Predicates: notNull, isSameAs(), in(), compose()
 * Sets: newSetFromMap(), newWeakHashSet(), newSoftHashSet(), newEnumSet()
 * Itera...s: size(), isEmpty(), limit(), skip(), getLast()
 * Iterators: emptyListIterator(); peekingIterator() (returning new
   PeekingIterator interface)
 * Ordering: like "Comparator++"; has max(), min(), sort(), sortedCopy(),
   binarySearch(), reverseOrder(); Comparators methods return these now
 * ObjectArrays: newArray(T[], int)
 * Suppliers: ofInstance(), compose(), memoize()
 * ImmutableList class (replaces Lists.immutableList())
 * ImmutableSet class (replaces Sets.immutableSet())

Lots of hard work done by Jared Levy, Jesse Wilson, Chris Povirk, Mick
Killianey, myself and others.

+++ src/com/google/common/collect/AbstractMapBasedMultiset.java
@@ -20,7 +20,6 @@ import com.google.common.base.Nullable;
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.base.Preconditions.checkState;
-
 import java.io.Serializable;
 import java.util.AbstractSet;
 import java.util.Collection;
@@ -139,7 +138,11 @@ abstract class AbstractMapBasedMultiset<E> extends AbstractMultiset<E>
     return new MapBasedMultisetIterator();
   }
 
-  // TODO: see if we can extend the superclass's iterator
+  /* 
+   * Not subclassing AbstractMultiset$MultisetIterator because next() needs to
+   * retrieve the Map.Entry<E, AtomicInteger> entry, which can then be used for
+   * a more efficient remove() call.
+   */
   private class MapBasedMultisetIterator implements Iterator<E> {
     final Iterator<Map.Entry<E, AtomicInteger>> entryIterator;
     Map.Entry<E, AtomicInteger> currentEntry;
@@ -204,7 +207,7 @@ abstract class AbstractMapBasedMultiset<E> extends AbstractMultiset<E>
       backingMap.put(element, new AtomicInteger(occurrences));
     } else {
       long newCount = (long) frequency.get() + (long) occurrences;
-      checkArgument(newCount <= (long) Integer.MAX_VALUE,
+      checkArgument(newCount <= Integer.MAX_VALUE,
           "too many occurrences: %s", newCount);
       frequency.getAndAdd(occurrences);
     }
